+++ exposed-core/src/main/kotlin/org/jetbrains/exposed/sql/SQLExpressionBuilder.kt
@@ -63,7 +63,7 @@ object SqlExpressionBuilder {
         return EqOp(this, wrap(t))
     }
 
-    infix fun <T:Comparable<T>> Column<EntityID<T>>.eq(t: T?) : Op<Boolean> {
+    infix fun <T:Comparable<T>> ExpressionWithColumnType<EntityID<T>>.eq(t: T?) : Op<Boolean> {
         if (t == null) {
             return isNull()
         }
@@ -80,7 +80,7 @@ object SqlExpressionBuilder {
         return NeqOp(this, wrap(other))
     }
 
-    infix fun <T:Comparable<T>> Column<EntityID<T>>.neq(t: T?) : Op<Boolean> {
+    infix fun <T:Comparable<T>> ExpressionWithColumnType<EntityID<T>>.neq(t: T?) : Op<Boolean> {
         if (t == null) {
             return isNotNull()
         }
@@ -95,17 +95,29 @@ object SqlExpressionBuilder {
 
     infix fun<T:Comparable<T>, S: T?> ExpressionWithColumnType<in S>.less(t: T) : Op<Boolean> = LessOp(this, wrap(t))
 
+    @JvmName("lessEntityID")
+    infix fun<T:Comparable<T>> ExpressionWithColumnType<EntityID<T>>.less(t: T) : Op<Boolean> = LessOp(this, wrap(t))
+
     fun<T:Comparable<T>, S: T?> ExpressionWithColumnType<in S>.less(other: Expression<S>) = LessOp(this, other)
 
     infix fun<T:Comparable<T>, S: T?> ExpressionWithColumnType<in S>.lessEq(t: T) : Op<Boolean> = LessEqOp(this, wrap(t))
+
+    @JvmName("lessEqEntityID")
+    infix fun<T:Comparable<T>> ExpressionWithColumnType<EntityID<T>>.lessEq(t: T) : Op<Boolean> = LessEqOp(this, wrap(t))
 
     fun<T:Comparable<T>, S: T?> ExpressionWithColumnType<in S>.lessEq(other: Expression<S>) : Op<Boolean> = LessEqOp(this, other)
 
     infix fun<T:Comparable<T>, S: T?> ExpressionWithColumnType<in S>.greater(t: T) : Op<Boolean> = GreaterOp(this, wrap(t))
 
+    @JvmName("greaterEntityID")
+    infix fun<T:Comparable<T>> ExpressionWithColumnType<EntityID<T>>.greater(t: T) : Op<Boolean> = GreaterOp(this, wrap(t))
+
     fun<T:Comparable<T>, S: T?> ExpressionWithColumnType<in S>.greater(other: Expression<S>) : Op<Boolean> = GreaterOp(this, other)
 
     infix fun<T:Comparable<T>, S: T?> ExpressionWithColumnType<in S>.greaterEq(t: T) : Op<Boolean> = GreaterEqOp(this, wrap(t))
+
+    @JvmName("greaterEqEntityID")
+    infix fun<T:Comparable<T>> ExpressionWithColumnType<EntityID<T>>.greaterEq(t: T) : Op<Boolean> = GreaterEqOp(this, wrap(t))
 
     fun<T:Comparable<T>, S: T?> ExpressionWithColumnType<in S>.greaterEq(other: Expression<T>) : Op<Boolean> = GreaterEqOp(this, other)