+++ exposed-core/src/main/kotlin/org/jetbrains/exposed/sql/vendors/OracleDialect.kt
@@ -6,9 +6,9 @@ import org.jetbrains.exposed.sql.transactions.TransactionManager
 
 internal object OracleDataTypeProvider : DataTypeProvider() {
 
-    override fun shortAutoincType() = "NUMBER(12)"
+    override fun integerAutoincType() = "NUMBER(12)"
 
-    override fun shortType() = "NUMBER(12)"
+    override fun integerType() = "NUMBER(12)"
 
     override fun longAutoincType() = "NUMBER(19)"