+++ exposed-core/src/main/kotlin/org/jetbrains/exposed/sql/vendors/SQLiteDialect.kt
@@ -5,7 +5,7 @@ import org.jetbrains.exposed.sql.*
 import org.jetbrains.exposed.sql.transactions.TransactionManager
 
 internal object SQLiteDataTypeProvider : DataTypeProvider() {
-    override fun shortAutoincType(): String = "INTEGER AUTOINCREMENT"
+    override fun integerAutoincType(): String = "INTEGER AUTOINCREMENT"
     override fun longAutoincType(): String = "INTEGER AUTOINCREMENT"
     override fun floatType(): String = "SINGLE"
     override fun booleanToStatementString(bool: Boolean) = if (bool) "1" else "0"